services/pricing/requirements.txt
# services/pricing/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg # O psycopg2-binary si prefieres PostgreSQL síncrono
python-dotenv
pydantic-settings
aio_pika # Para RabbitMQ si lo usas
httpx # Para llamadas a otros servicios si es necesario

# Dependencias del Scraper
selenium
webdriver-manager
requests # Útil para algunas tareas
# lxml # A menudo útil para parsear HTML si no usas regex/CSS selectors